+++ test/native/test_main.cpp
@@ -2,6 +2,7 @@
 
 #include "test_framework.h"
 
+void runAudioFeedbackTests(TestRunner& runner);
 void runCoinTests(TestRunner& runner);
 void runCoinAnimationTests(TestRunner& runner);
 void runDivinationFlowTests(TestRunner& runner);
@@ -12,6 +13,7 @@ void runMicrophoneResearchTests(TestRunner& runner);
 void runPowerControllerTests(TestRunner& runner);
 void runPowerManagerTests(TestRunner& runner);
 void runTransformTests(TestRunner& runner);
+void runUiLayoutTests(TestRunner& runner);
 void runOrderingTests(TestRunner& runner);
 void runStateMachineTests(TestRunner& runner);
 void runHistoryRecordTests(TestRunner& runner);
@@ -22,6 +24,7 @@ void runWifiTests(TestRunner& runner);
 
 int main() {
   TestRunner runner;
+  runAudioFeedbackTests(runner);
   runCoinTests(runner);
   runCoinAnimationTests(runner);
   runDivinationFlowTests(runner);
@@ -32,6 +35,7 @@ int main() {
   runPowerControllerTests(runner);
   runPowerManagerTests(runner);
   runTransformTests(runner);
+  runUiLayoutTests(runner);
   runOrderingTests(runner);
   runStateMachineTests(runner);
   runHistoryRecordTests(runner);